requirements.txt
Flask==2.0.1
Werkzeug==2.0.1
google-generativeai==0.1.0